remove unnecessary BaseEventBroadcaster and rework things
